Init work of version 2
