--- docs/assets/css/index.90df3929.chunk.css
@@ -1,1 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Oxygen&display=swap);.month-picker__container{width:400px;position:relative;box-shadow:inset 0 0 0 1px rgba(0,0,0,.1),inset 0 -1px 0 rgba(0,0,0,.1);border-radius:5px}.month-picker__container.year-picker{width:20rem}.month-picker{box-sizing:border-box;flex:1;width:auto;display:flex;flex-direction:row;flex-wrap:wrap;box-shadow:1px 2px 5px rgba(0,0,0,.1);font-family:sans-serif;border-radius:5px;overflow:hidden;display:grid;grid-template-columns:1fr 1fr 1fr}.month-picker__year{padding:.2rem;background-color:#fcfcfc;box-shadow:inset 0 0 0 1px rgba(0,0,0,.1),inset 0 -1px 0 rgba(0,0,0,.1)}.month-picker__year p{width:100%;font-weight:600;letter-spacing:1px;font-size:1.2rem;margin:1rem 0}.month-picker__year input{padding:0;font-weight:600;border-radius:5px 5px 0 0;outline:none;border:none;font-size:1.2rem;width:auto;text-align:center;box-sizing:border-box;width:100%;height:3.5rem;position:relative;z-index:1}.month-picker__year input:focus{border:1px solid #55b0f2}.month-picker__year button,.month-picker__year div,.month-picker__year p{text-align:center;flex:1}.month-picker__year button{background-color:#f4f4f4;position:absolute;width:5rem;height:2.75rem;font-size:2rem;border-radius:5px;outline:none;top:.5rem;border:1px solid #e8e8e8;z-index:2;color:#686868}.month-picker__year button:hover{background-color:rgba(0,0,0,.025)}.month-picker__year button:active{background-color:rgba(0,0,0,.04)}.month-picker__year button:first-child{left:10px}.month-picker__year button:last-child{right:10px}.month-picker__month{flex-basis:calc(33.333% - 10px);padding:.75rem .25rem;cursor:pointer;text-align:center;border:1px solid hsla(0,0%,96.1%,.75);transition:all .25s cubic-bezier(.165,.84,.44,1);background-color:#fefefe;user-select:none;position:relative}.month-picker__month:hover:after{display:block;content:"";position:absolute;width:95%;height:95%;top:50%;left:50%;transform:translate(-50%,-50%);box-shadow:0 0 10px rgba(0,0,0,.1);z-index:10}.month-picker__month.selected,.month-picker__month.selected-range-first,.month-picker__month.selected-range-second{background-color:#55b0f2;color:#fff;border-radius:5px;box-shadow:inset 0 0 3px #3490d2,0 2px 5px rgba(85,176,242,.2);text-shadow:0 2px 4px rgba(0,0,0,.25)}.month-picker__month.selected-range-first,.month-picker__month.selected-range-second{border-color:transparent}.month-picker__month.selected-range{background-color:#7eb9e2;color:#fff;border-color:transparent}.month-picker__month.inactive{background-color:#f0f0f0;color:#8a8a8a;cursor:not-allowed;box-shadow:none;text-shadow:none}.month-picker__month.inactive:hover{box-shadow:none}.month-picker--dark{background-color:#5f5f5f}.month-picker--dark .month-picker__year{background-color:#2f2f30}.month-picker--dark .month-picker__year input,.month-picker--dark .month-picker__year p{color:#ebebeb}.month-picker--dark .month-picker__year input{background-color:#5f5f5f}.month-picker--dark .month-picker__year button{background-color:#505050;color:#c9c9c9;border-color:#1e1e1e}.month-picker--dark .month-picker__year button:hover{background-color:rgba(0,0,0,.3)}.month-picker--dark .month-picker__year button:active{background-color:rgba(0,0,0,.6)}.month-picker--dark .month-picker__month{background-color:#2f2f30;border-color:hsla(0,0%,96.1%,.15);color:#c9c9c9}.month-picker--dark .month-picker__month.selected,.month-picker--dark .month-picker__month.selected-range-first,.month-picker--dark .month-picker__month.selected-range-second{background-color:#505050;box-shadow:inset 0 0 3px #505050,0 2px 5px #505050;color:#fff;border-color:#1d1b1b}.month-picker--dark .month-picker__month.selected-range{background-color:#636363;color:#fff;border-color:transparent}.month-picker--dark .month-picker__month:hover{box-shadow:0 0 6px rgba(0,0,0,.8);z-index:10}.month-picker--dark .month-picker__month.inactive{background-color:#3f3f3f;color:#8a8a8a;cursor:not-allowed;box-shadow:none;text-shadow:none}@media only screen and (max-width:768px){.month-picker__container{width:100%}}.month-picker-input-container[data-v-4455fe98]{position:relative;width:180px;min-width:140px}.month-picker-input[data-v-4455fe98]{padding:1em 1.5em;font-size:.85em;border-radius:5px;outline:none;border:1px solid rgba(0,0,0,.15);transition:all .35s cubic-bezier(.165,.84,.44,1)}.month-picker-input[data-v-4455fe98]:focus{border-color:rgba(0,0,0,.25);box-shadow:0 1px 2px rgba(0,0,0,.1)}.month-picker__container[data-v-4455fe98]{position:absolute;top:3.5em}:root{background-color:#efefef}.content[data-v-4d992784]{max-width:800px;margin:0 auto;font-family:Oxygen,sans-serif;padding-bottom:20rem}.content__grid[data-v-4d992784]{display:grid;grid-template-columns:.5fr 1fr;gap:4rem}.content h1[data-v-4d992784]{margin-bottom:2em;letter-spacing:2px;text-shadow:2px 4px 10px #e4e4e4}.github-corner:hover .octo-arm[data-v-4d992784]{animation:octocat-wave-4d992784 .56s ease-in-out}.form__container[data-v-4d992784]{padding:.25rem 0}.form__label[data-v-4d992784]{display:grid;grid-template-columns:3fr 2fr;width:20rem;padding:.25rem;border-radius:5px}.form__label input[type=checkbox][data-v-4d992784]{width:1rem;height:1rem}.form__label[data-v-4d992784]:hover{background-color:#f7f7f7}.form__label select[data-v-4d992784]{padding:0 2rem}pre[data-v-4d992784]{font-family:monospace;font-size:1rem;font-weight:600;color:#313131}@keyframes octocat-wave-4d992784{0%{transform:rotate(0deg)}20%{transform:rotate(-25deg)}40%{transform:rotate(10deg)}60%{transform:rotate(-25deg)}80%{transform:rotate(10deg)}to{transform:rotate(0deg)}}@media (max-width:500px){.github-corner:hover .octo-arm[data-v-4d992784]{animation:none}.github-corner .octo-arm[data-v-4d992784]{animation:octocat-wave-4d992784 .56s ease-in-out}}
