--- docs/assets/js/chunk-vendors.69cce7e2.chunk.js.LICENSE.txt
@@ -1,5 +1,0 @@
-/*!
- * Vue.js v2.7.10
- * (c) 2014-2022 Evan You
- * Released under the MIT License.
- */
