--- docs/index.html
@@ -1,129 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width,initial-scale=1" />
-        <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-        <title>vue-month-picker</title>
-        <meta
-            name="description"
-            content="A simple and elegant month picker for Vue"
-        />
-        <link href="./assets/css/index.90df3929.chunk.css" rel="stylesheet" />
-    </head>
-    <body>
-        <noscript
-            >We're sorry but vue-month-picker doesn't work properly without
-            JavaScript enabled. Please enable it to continue.</noscript
-        >
-        <div id="app"></div>
-        <script>
-            !(function (e) {
-                function r(r) {
-                    for (
-                        var n, l, f = r[0], i = r[1], a = r[2], c = 0, s = [];
-                        c < f.length;
-                        c++
-                    )
-                        (l = f[c]),
-                            Object.prototype.hasOwnProperty.call(o, l) &&
-                                o[l] &&
-                                s.push(o[l][0]),
-                            (o[l] = 0)
-                    for (n in i)
-                        Object.prototype.hasOwnProperty.call(i, n) &&
-                            (e[n] = i[n])
-                    for (p && p(r); s.length; ) s.shift()()
-                    return u.push.apply(u, a || []), t()
-                }
-                function t() {
-                    for (var e, r = 0; r < u.length; r++) {
-                        for (var t = u[r], n = !0, f = 1; f < t.length; f++) {
-                            var i = t[f]
-                            0 !== o[i] && (n = !1)
-                        }
-                        n && (u.splice(r--, 1), (e = l((l.s = t[0]))))
-                    }
-                    return e
-                }
-                var n = {},
-                    o = { 2: 0 },
-                    u = []
-                function l(r) {
-                    if (n[r]) return n[r].exports
-                    var t = (n[r] = { i: r, l: !1, exports: {} })
-                    return (
-                        e[r].call(t.exports, t, t.exports, l),
-                        (t.l = !0),
-                        t.exports
-                    )
-                }
-                ;(l.m = e),
-                    (l.c = n),
-                    (l.d = function (e, r, t) {
-                        l.o(e, r) ||
-                            Object.defineProperty(e, r, {
-                                enumerable: !0,
-                                get: t,
-                            })
-                    }),
-                    (l.r = function (e) {
-                        'undefined' != typeof Symbol &&
-                            Symbol.toStringTag &&
-                            Object.defineProperty(e, Symbol.toStringTag, {
-                                value: 'Module',
-                            }),
-                            Object.defineProperty(e, '__esModule', {
-                                value: !0,
-                            })
-                    }),
-                    (l.t = function (e, r) {
-                        if ((1 & r && (e = l(e)), 8 & r)) return e
-                        if (4 & r && 'object' == typeof e && e && e.__esModule)
-                            return e
-                        var t = Object.create(null)
-                        if (
-                            (l.r(t),
-                            Object.defineProperty(t, 'default', {
-                                enumerable: !0,
-                                value: e,
-                            }),
-                            2 & r && 'string' != typeof e)
-                        )
-                            for (var n in e)
-                                l.d(
-                                    t,
-                                    n,
-                                    function (r) {
-                                        return e[r]
-                                    }.bind(null, n)
-                                )
-                        return t
-                    }),
-                    (l.n = function (e) {
-                        var r =
-                            e && e.__esModule
-                                ? function () {
-                                      return e.default
-                                  }
-                                : function () {
-                                      return e
-                                  }
-                        return l.d(r, 'a', r), r
-                    }),
-                    (l.o = function (e, r) {
-                        return Object.prototype.hasOwnProperty.call(e, r)
-                    }),
-                    (l.p = '/')
-                var f = (window.webpackJsonp = window.webpackJsonp || []),
-                    i = f.push.bind(f)
-                ;(f.push = r), (f = f.slice())
-                for (var a = 0; a < f.length; a++) r(f[a])
-                var p = i
-                t()
-            })([])
-        </script>
-        <script src="./assets/js/chunk-vendors.69cce7e2.chunk.js"></script>
-        <script src="./assets/js/index.90df3929.chunk.js"></script>
-    </body>
-</html>
