--- poi.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    entry: 'example/index',
-    plugins: [
-        {
-            resolve: '@poi/plugin-eslint',
-        },
-    ],
-}
