+++ src/App.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import MonthPicker from './components/MonthPicker.vue'
+</script>
+
+<template>
+    <MonthPicker />
+</template>
+
+<style scoped></style>
